+++ sv-core/src/Data/Sv/Decode/Error.hs
@@ -1,3 +1,6 @@
+{-# language OverloadedStrings #-}
+{-# language ScopedTypeVariables #-}
+
 {-|
 Module      : Data.Sv.Decode.Error
 Copyright   : (C) CSIRO 2017-2019
@@ -22,6 +25,12 @@ module Data.Sv.Decode.Error (
 , badParse
 , badDecode
 
+-- * Display
+, displayErrors
+, displayErrors'
+, dieOnError
+, dieOnError'
+
 -- * Conversions
 , validateEither
 , validateEitherWith
@@ -32,9 +41,22 @@ module Data.Sv.Decode.Error (
 , bindValidation
 ) where
 
+import Data.ByteString (ByteString)
+import qualified Data.ByteString.Char8 as Char8
+import Data.Foldable (toList)
+import Data.List (intersperse)
+import Data.Semigroup (Semigroup ((<>)))
+import Data.Semigroup.Foldable (Foldable1 (foldMap1))
+import Data.String (IsString)
+import qualified Data.Text.Encoding as T
+import qualified Data.Text.Lazy as LT
+import qualified Data.Text.Lazy.IO as LT
+import Data.Text.Lazy.Builder (Builder)
+import qualified Data.Text.Lazy.Builder as Builder
 import Data.Validation (Validation (Failure), bindValidation)
 import Data.Vector (Vector)
 import qualified Text.Trifecta.Result as Trifecta
+import System.Exit (exitFailure)
 
 import Data.Sv.Decode.Type
 
@@ -107,3 +129,77 @@ validateTrifectaResult f =
       trifectaResultToEither r = case r of
         Trifecta.Failure e -> Left . show . Trifecta._errDoc $ e
         Trifecta.Success a -> Right a
+
+-- | Pretty print errors as a string. Each error is on its own line.
+displayErrors :: DecodeErrors ByteString -> LT.Text
+displayErrors = displayErrors' buildBytestring
+
+-- | Pretty print errors as a string. Each error is on its own line.
+--
+-- This version lets you work with any 'String' type in your errors.
+displayErrors' :: forall e. (e -> Builder) -> DecodeErrors e -> LT.Text
+displayErrors' build (DecodeErrors errs) =
+  let
+    indent :: Builder -> Builder
+    indent x = "  " <> x
+
+    displayErr :: DecodeError e -> Builder
+    displayErr e = indent $ case e of
+      BadParse msg -> "Parsing the document failed. The error was: " <> build msg
+      UnexpectedEndOfRow -> "Expected more fields, but the row ended."
+      ExpectedEndOfRow extras ->
+        "Expected fewer fields in the row. The extra fields contained: " <>
+          commaSep (bquote <$> toList extras)
+      UnknownCategoricalValue found required ->
+        "Unknown categorical value found: " <> bquote found <> ". Expected one of: " <>
+          (commaSep . fmap bquote . mconcat) required
+      MissingColumn name -> "Could not find required column " <> bquote name
+      MissingHeader -> "A header row was required, but one was not found."
+      BadConfig msg -> "sv was misconfigured: " <> build msg
+      BadDecode msg -> "Decoding a field failed: " <> build msg
+
+    displayAndCount = count . displayErr
+    Counted body c = foldMap1 displayAndCount errs
+    spaceSep = mconcat . intersperse " "
+    commaSep = mconcat . intersperse ", "
+    quote s = "\"" <> s <> "\""
+    bquote = quote . build
+    pluralise n s =
+      if n == 1
+      then s
+      else Builder.fromString (show n) <> " " <> s <> "s"
+    heading = spaceSep ["The following", pluralise c "error", "occurred:"]
+  in
+    Builder.toLazyText $ heading <> "\n" <> body
+
+-- | If the 'DecodeValidation' is a 'Failure', print a pretty error message
+-- and call 'exitFailure'
+dieOnError :: DecodeValidation ByteString a -> IO a
+dieOnError = dieOnError' buildBytestring
+
+-- | If the 'DecodeValidation' is a 'Failure', print a pretty error message
+-- and call 'exitFailure'
+--
+-- This version lets you work with different String types.
+dieOnError' :: (e -> Builder) -> DecodeValidation e a -> IO a
+dieOnError' build e = case e of
+  Failure errs -> do
+    LT.putStrLn $ displayErrors' build errs
+    exitFailure
+  Success a -> pure a
+
+---- internal
+
+buildBytestring :: ByteString -> Builder
+buildBytestring bs = case T.decodeUtf8' bs of
+  Left  _ -> Builder.fromString $ Char8.unpack bs
+  Right b -> Builder.fromText b
+
+data Counted e = Counted e Integer
+
+count :: e -> Counted e
+count e = Counted e 1
+
+instance (Semigroup e, IsString e) => Semigroup (Counted e) where
+  Counted b c <> Counted b' c' =
+    Counted (b <> "\n" <> b') (c+c')